+++ R/future-invoke-map.R
@@ -159,6 +159,31 @@ future_invoke_map_lgl <- function(.f,
 
 #' @rdname future_invoke_map
 #' @export
+future_invoke_map_raw <- function(.f,
+                                  .x = list(NULL),
+                                  ...,
+                                  .env = NULL,
+                                  .options = furrr_options(),
+                                  .env_globals = parent.frame(),
+                                  .progress = deprecated()) {
+  maybe_warn_deprecated_progress(is_present(.progress), what = "future_invoke_map_raw")
+
+  .env <- .env %||% parent.frame()
+  .f <- as_invoke_function(.f)
+
+  future_map2_raw(
+    .x = .f,
+    .y = .x,
+    .f = purrr::invoke,
+    ...,
+    .env = .env,
+    .options = .options,
+    .env_globals = .env_globals
+  )
+}
+
+#' @rdname future_invoke_map
+#' @export
 future_invoke_map_dfr <- function(.f,
                                   .x = list(NULL),
                                   ...,