+++ tests/testthat/test-future-pmap.R
@@ -62,6 +62,16 @@ furrr_test_that("future_pmap_chr() works", {
   )
 })
 
+furrr_test_that("future_pmap_raw() works", {
+  x <- c("a", "b", "c")
+  y <- as.raw(1:3)
+
+  expect_identical(
+    future_pmap_raw(list(x, y), ~.y),
+    pmap_raw(list(x, y), ~.y)
+  )
+})
+
 furrr_test_that("names of `.x` are retained", {
   x <- c(a = 1, b = 2)
   y <- c(c = 1, d = 2)
@@ -108,6 +118,7 @@ furrr_test_that("atomic variants work with size zero input", {
   expect_identical(future_pmap_dbl(list(list(), list()), identity), double())
   expect_identical(future_pmap_int(list(list(), list()), identity), integer())
   expect_identical(future_pmap_lgl(list(list(), list()), identity), logical())
+  expect_identical(future_pmap_raw(list(list(), list()), identity), raw())
 })
 
 furrr_test_that("size one recycling works", {